Delete package-lock.json file
The package-lock.json file has been deleted from the project. This file is automatically generated for any operations where npm modifies either the node_modules tree, or package.json. It describes the exact tree that was generated, such that subsequent installs are able to generate identical trees, regardless of intermediate dependency updates. This file is not necessary for the project and was, therefore, removed.