Add package-lock.json file with dependencies
A new package-lock.json file has been added, containing a detailed account of packages and their precise version numbers at the time of installation. The file ensures that the exact set of dependencies specified during installation will be consistent across all environments. This guarantees the proper functioning of the project irrespective of the version of npm that's being used, and handles updates of any intermediate dependencies in an efficient manner.